// src/lib/zod-schema/problem-input-schema.ts
import { z } from "zod";

export const ProblemInputSchema = z.object({
  id: z.string().min(1, "ID is required"), // only temporary, will be generated by the backend
  title: z
    .string()
    .min(1, "Title is required")
    .max(200, "Title must be less than 200 characters"),
  description: z.string().min(1, "Description is required"),
  imageUrl: z.string().url("Must be a valid URL").nullable().optional(),
  classroomId: z.string().nullable().optional(),
  isPublic: z.boolean().default(true),
  createdAt: z.date(), // only temporary, will be generated by the backend
});

export type ProblemInput = z.infer<typeof ProblemInputSchema>;